refactor(attn): move the page-unit geometry into the package for it
`page_unit_geometry` landed on main while this branch was splitting the flat
`attentions/` directory into packages named for the question each answers. It
belongs in `pool_layout/`, and says so itself: its docstring opens "Where a
checkpoint image's bytes live in the MLA paged pool" -- the package's topic --
and justifies its own existence by being reachable without aiter so the
geometry deciding where every checkpoint byte goes can be tested on a GPU-less
runner, which is the package's invariant. It imports neither aiter nor atom.

The module no longer restates that invariant; `pool_layout/__init__.py` is
where it is written, and `tests/test_layout_packages.py` now enforces it over
six members rather than five.
