+++ atom/model_ops/attentions/kimi_mla_gdn_attn.py
@@ -24,7 +24,7 @@ from .aiter_mla import (
 )
 from .backends import AttentionBackend
 from .gdn_attn import GDNStateMixin
-from .page_unit_geometry import PageUnitGeometryMixin
+from .pool_layout.page_unit_geometry import PageUnitGeometryMixin
 from .pool_layout.sub_pool_spec import SubPoolSpec, page_pool, state_pool
 from .triton_mla import TritonMLAMetadataBuilder
